% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestree.R
\name{ancestree}
\alias{ancestree}
\title{ancestree: Get a Phylogeny that Starts at the Common Ancestor of a Group of Species}
\usage{
ancestree(sample, tree)
}
\arguments{
\item{sample}{A data frame with species in columns' names (check out example link).}

\item{tree}{.tre phylogenetic tree of species.}
}
\value{
A phylogeny that starts at the common ancestor of the provided sample.
}
\description{
Uses a sample of species to get the phylogeny that starts at the common ancestor of the sample.
}
\examples{
# https://rfunctions.blogspot.com/2013/04/ancestree-function-get-phylogeny-that.html
}
