% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astRaRium.R
\name{astRaRium}
\alias{astRaRium}
\title{astRaRium: Generate infinite Astrarium minigames.}
\usage{
astRaRium()
}
\value{
You play the game. You can succeed or fail.
}
\description{
Generate infinite minigames based on the Astrarium minigame, from "Dragon Age: Inquisition". You have to click on the stars (asterisks) to find a path that that reveals the same image as shown while not repeating connections already made once. When it begins (by typing "astRaRium()" ), it will ask you for the desired number of stars and the number of links between pairs of stars. A good number is 12 and 9, respectively.
}
\examples{
# https://rfunctions.blogspot.com/2021/05/astrarium2-r-game-about-linking-stars.html
}
