% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.map.R
\name{bivariate.map}
\alias{bivariate.map}
\title{bivariate.map: Create a Bivariate Map}
\usage{
bivariate.map(
  rasterx,
  rastery,
  colormatrix,
  nquantiles = 10,
  trivariate = F,
  rasterz
)
}
\arguments{
\item{rasterx}{raster}

\item{rastery}{raster}

\item{colormatrix}{color matrix from colmat() function}

\item{nquantiles}{number of quantiles in color matrix (same as used when using colmat() function)}

\item{trivariate}{boolean indicating if should use a third raster for a "trivariate" map where the third variable corresponds to the transparency (alpha) of colors.}

\item{rasterz}{raster (only if trivariate=T)}
}
\value{
A plot with the bivariate map.
}
\description{
Creates a Bivariate Map using two rasters and a color matrix created with colmat() function.
}
\examples{
# https://rfunctions.blogspot.com/2015/03/bivariate-maps-bivariatemap-function.html
}
