% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridFilter.R
\name{GridFilter}
\alias{GridFilter}
\title{GridFilter: Intersect Shape with a Grid and Exclude Cells Based on Area Coverage}
\usage{
GridFilter(shape, resol = 1, prop = 0)
}
\arguments{
\item{shape}{shapefile}

\item{resol}{resolution}

\item{prop}{minimum value of area covered by the grid cell. The default is 0 (i.e. it does not delete any grid cell)}
}
\value{
A "gridded" shapefile. Plot this output to take a look at it.
}
\description{
Creates a shape intersected with a grid. The user can exclude cells based on area coverage. For example, if the shape covers only 50 percent of some cells, the user can choose to exclude or maintain these cells.
}
\examples{
# https://rfunctions.blogspot.com/2014/12/gridfilter-intersect-grid-with-shape.html
}
